% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARBOLhelper.R
\name{ARBOL.Aggregate}
\alias{ARBOL.Aggregate}
\title{ARBOL Aggregate Function}
\usage{
ARBOL.Aggregate(node, attribute, aggFun, ...)
}
\arguments{
\item{node}{A node in the tree. Can be either a leaf or a parent node.}

\item{attribute}{The attribute that should be used for aggregation.}

\item{aggFun}{The function used for aggregating the attribute values. This function should take a vector and return a single value.}

\item{...}{Additional arguments to pass to GetAttribute().}
}
\value{
Returns the aggregated attribute value.
}
\description{
This function recursively traverses a node tree to aggregate attribute values.
}
\note{
The 'cacheAttribute' argument is deprecated. Please use 'Do' instead.
}
\examples{
# Assuming 'treeRoot' is the root of your tree and 'sum' is your aggregation function
ARBOL.Aggregate(treeRoot, "height", sum)

}
