% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARBOLhelper.R
\name{as.dendrogram.NodePS}
\alias{as.dendrogram.NodePS}
\title{data.tree to ggraph object conversion with custom Node -> dend object conversion function that uses pathString as node names.}
\usage{
\method{as.dendrogram}{NodePS}(
  object,
  heightAttribute = "plotHeight",
  edgetext = FALSE,
  namefield = "pathString",
  ...
)
}
\arguments{
\item{object}{A data tree object}

\item{heightAttribute}{plot height attribute. Default is plotHeight}
}
\value{
dendrogram object
}
\description{
data.tree to ggraph object conversion with custom Node -> dend object conversion function that uses pathString as node names.
}
\examples{
txt <- ToNewickPS(divtree)
ARBOLphylo <- ape::read.tree(text=txt)
}
