% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARBOLhelper.R
\name{as.phylo.NodeI}
\alias{as.phylo.NodeI}
\title{Fast Conversion of data.tree Node to ape Phylo Object}
\usage{
\method{as.phylo}{NodeI}(x, heightAttribute = "plotHeight")
}
\arguments{
\item{x}{A data.tree node to be converted.}

\item{heightAttribute}{The attribute used to define the height of the nodes
in the resulting phylo object. Defaults to 'plotHeight'.}
}
\value{
A phylo object representing the phylogenetic tree, with internal nodes
        named by their `pathString` and leaf nodes named by their end name.
}
\description{
Efficiently converts a data.tree node to a phylogenetic tree (`phylo`) object
from the `ape` package via manipulation of an intermediate dendrogram object.
Slashes from data.tree must be replaced with .
in I version, internal nodes are named with pathString but leaf nodes have only leaf name
}
\examples{
# Assuming `myTree` is a data.tree Node
phylo_tree <- as.phylo.NodeI(myTree)
plot(phylo_tree)
}
