% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARBOLhelper.R
\name{as.phylo.NodeNW}
\alias{as.phylo.NodeNW}
\title{Convert data.tree Node to ape phylo Object through Newick Format
This function converts a data.tree node to a phylogenetic tree (`phylo`)
object from the `ape` package by first converting the data.tree to a Newick
string format. This method is slower compared to direct conversion methods,
but may be useful if ape or data.tree make updates that make other methods 
incompatible}
\usage{
\method{as.phylo}{NodeNW}(x, heightAttribute = DefaultPlotHeight, ...)
}
\arguments{
\item{x}{A data.tree node to be converted.}

\item{heightAttribute}{The attribute used to define the height of the nodes
in the resulting phylo object. Defaults to 'DefaultPlotHeight'.}

\item{...}{Additional arguments passed to `ape::read.tree`.}
}
\value{
A phylo object representing the phylogenetic tree.
}
\description{
Convert data.tree Node to ape phylo Object through Newick Format
This function converts a data.tree node to a phylogenetic tree (`phylo`)
object from the `ape` package by first converting the data.tree to a Newick
string format. This method is slower compared to direct conversion methods,
but may be useful if ape or data.tree make updates that make other methods 
incompatible
}
\examples{
# Assuming `myTree` is a data.tree Node
phylo_tree <- as.phylo.NodeNW(myTree)
plot(phylo_tree)
}
