% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCTieredClustering.R
\name{BaseCondition_default}
\alias{BaseCondition_default}
\title{How to determine of tiered clustering should stop}
\usage{
BaseCondition_default(srobj, min_cluster_size = 100, max_tiers = 10)
}
\arguments{
\item{srobj}{v3 seurat object}

\item{min_cluster_size}{defaults to 100}

\item{max_tiers}{defaults to 10, determines last tier}
}
\value{
0 if should continue, ints > 0 indicate first base condition reached
}
\description{
How to determine of tiered clustering should stop
}
