% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARBOLhelper.R
\name{binaryTreeToDF}
\alias{binaryTreeToDF}
\title{Converts pvclust tree object to dataframe with row per node and pvclust p-values}
\usage{
binaryTreeToDF(pvclust_tree)
}
\arguments{
\item{pvclust_tree}{a pvclust or hclust tree}
}
\value{
a dataframe with one row per node of the tree
}
\description{
Converts pvclust tree object to dataframe with row per node and pvclust p-values
}
\examples{
binarydf <- binaryTreeToDF(pvclust_tree=srobj@misc$pvclust)
}
