% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCTieredClustering.R
\name{binomcount.test}
\alias{binomcount.test}
\title{Binomial signifcance test for on/off gene expression}
\usage{
binomcount.test(object, cells.1, cells.2, effect.size, TPM.mat, Count.mat)
}
\arguments{
\item{object}{Seruat object}

\item{cells.1}{cell names of group 1}

\item{cells.2}{cell names of group 2}

\item{effect.size}{log2 foldchange threshold for effect size significance}

\item{TPM.mat}{Transcripts/million matrix}

\item{Count.mat}{Count matrix}
}
\value{
significant gene data.frame
}
\description{
Binomial signifcance test for on/off gene expression
}
