% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARBOLhelper.R
\name{bintree_to_df}
\alias{bintree_to_df}
\title{Converts pvclust tree object to dataframe with row per node
calls prepARBOLmeta_tree() and prepTree()}
\usage{
bintree_to_df(pvclust_tree)
}
\arguments{
\item{pvclust_tree}{a pvclust or hclust tree}
}
\value{
a dataframe with one row per node of the tree
}
\description{
Converts pvclust tree object to dataframe with row per node
calls prepARBOLmeta_tree() and prepTree()
}
\examples{
binarydf <- bintree_to_df(pvclust_tree=srobj@misc$pvclust)
}
