% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARBOL.R, R/arboldotdotdot.R
\name{Choose_harmony_vectors}
\alias{Choose_harmony_vectors}
\title{always use elbow plot for harmony choice as jackstraw is not supported with reductions other than PCA
  if none sig use first 2 pcs}
\usage{
Choose_harmony_vectors(
  srobj,
  improved_diff_quantile = 0.85,
  significance = 0.01,
  figure_dir = NULL
)

Choose_harmony_vectors(
  srobj,
  improved_diff_quantile = 0.85,
  significance = 0.01,
  figure_dir = NULL
)
}
\description{
always use elbow plot for harmony choice as jackstraw is not supported with reductions other than PCA
  if none sig use first 2 pcs

always use elbow plot for harmony choice as jackstraw is not supported with reductions other than PCA
  if none sig use first 2 pcs
}
