% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCTieredClustering.R
\name{ChooseClusterResolutionDownsample}
\alias{ChooseClusterResolutionDownsample}
\title{Chooses Optimal Resolution for louvain clustering based on Silhouette score}
\usage{
ChooseClusterResolutionDownsample(
  input.srobj,
  assay = "RNA",
  n.pcs = input.srobj@misc$nPCs,
  sample.name = format(Sys.time(), "\%a-\%b-\%d-\%X-\%Y-\%Z"),
  res.low = 0.01,
  res.high = 3,
  res.n = 40,
  bias = "under",
  figdir = NULL
)
}
\arguments{
\item{input.srobj}{v3 seurat object}

\item{assay}{assay to cluster on in Seurat object}

\item{n.pcs}{which PCs to use (1:n) defaults to `input.srobj@misc$nPC` which needs to be set by user if used outside of tiered clustering}

\item{sample.name}{name of sample defaults to Sys.time}

\item{res.low}{inclusive lower bound of resolutions}

\item{res.high}{inclusive upper bound of resolutions}

\item{res.n}{how many resolutions to try within range}

\item{bias}{If multiple equal resolutions, choose fewer or more clusters with lower or higher resolution}

\item{figdir}{where to save QC plot}
}
\value{
seurat object with optimum clustering saved in `srobj$Best.Clusters`
}
\description{
Credit: Carly Ziegler
}
\details{
clusters across range of resolutions
selects random downsample of cells (10% of cells)
computes silhouette score for cells
chooses resolution with max avg mean silhouette score
}
