% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCTieredClustering.R
\name{ChooseHVG_default}
\alias{ChooseHVG_default}
\title{wrapper for basic seurat choosing}
\usage{
ChooseHVG_default(srobj)
}
\arguments{
\item{srobj}{seurat object}
}
\value{
srobj with highly variable features selected
}
\description{
runs: FindVariableFeatures(srobj, nfeatures=2000, selection.method = "disp")
}
