% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCTieredClustering.R
\name{ChooseOptimalClustering_default}
\alias{ChooseOptimalClustering_default}
\title{get resolution from paramsweep of heavily downsampled dataset}
\usage{
ChooseOptimalClustering_default(
  srobj,
  downsample_num = Inf,
  PreProcess_fun = PreProcess,
  ...
)
}
\description{
get resolution from paramsweep of heavily downsampled dataset
}
