% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARBOL.R, R/arboldotdotdot.R
\name{ChoosePCs_default}
\alias{ChoosePCs_default}
\title{Chooses the number of principle components to include in clustering}
\usage{
ChoosePCs_default(
  srobj,
  improved_diff_quantile = 0.85,
  significance = 0.01,
  figure_dir = NULL
)

ChoosePCs_default(
  srobj,
  improved_diff_quantile = 0.85,
  significance = 0.01,
  figure_dir = NULL
)
}
\arguments{
\item{srobj}{v4 seurat object}

\item{improved_diff_quantile}{percent variance explained of next PC to choose PC}

\item{significance}{JackStraw significance required to choose PC}

\item{figure_dir}{directory where PC choice plots are saved}
}
\value{
list of PCs chosen

list of PCs chosen
}
\description{
Seurat's JackStraw is used when there are less than 500 cells in a clustering event. 
otherwise, PCs are used in clustering if they explain 15% more variance than the following PC.

Also saves plots describing PC choice

Seurat's JackStraw is used when there are less than 500 cells in a clustering event. 
otherwise, PCs are used in clustering if they explain 15% more variance than the following PC.

Also saves plots describing PC choice
}
\examples{
nPCs <- ChoosePCs_default(srobj, figure_dir=fig_dir)

nPCs <- ChoosePCs_default(srobj, figure_dir=fig_dir)

}
