% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARBOL.R
\name{chooseResolution_SilhouetteAnalysisParameterScan}
\alias{chooseResolution_SilhouetteAnalysisParameterScan}
\title{Performs a parameter scan based on silhouette analysis on a seurat object}
\usage{
chooseResolution_SilhouetteAnalysisParameterScan(
  input.srobj,
  assay = cluster_assay,
  n.pcs = input.srobj@misc$nPCs,
  sample.name = format(Sys.time(), "\%a-\%b-\%d-\%X-\%Y-\%Z"),
  res.low = 0.01,
  res.high = 3,
  res.n = 40,
  bias = "under",
  figdir = NULL
)
}
\arguments{
\item{input.srobj}{v4 seurat object}

\item{assay}{seurat object assay to use for silhouette analysis}

\item{n.pcs}{number of PCs to include in clustering}

\item{sample.name}{names resulting parameter scan results graphs}

\item{res.low}{lowest resolution to include in parameter scan}

\item{res.high}{highest resolution}

\item{res.n}{number of resolutions to test}

\item{bias}{bias toward resolution choice, when resolution results are similar}

\item{figdir}{directory to save results graphs}
}
\value{
seurat object with slot misc$resolution.choice
}
\description{
Silhouette Analysis, a parameter scan method for choosing the most distinct clusters, is performed across 40 possible resolution on a seurat object. In ARBOL, srobj is downsampled for input to the parameter scan when there are >7.5k cells

Also saves plots describing the parameter scan results
}
\examples{
srobj <- readRDS("/path/to/seurat_object.rds")
srobj <- chooseResolution_SilhouetteAnalysisParameterScan(srobj)

}
