% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARBOLhelper.R
\name{data.tree_to_ggraphNW}
\alias{data.tree_to_ggraphNW}
\title{data.tree to ggraph conversion, by converting data.tree structure to ape via newick (doesn't carry annotations)
then ape (still doesn't carry annotations) to ggraph, 
then join data frame of data.tree to restore annotations.
requires 'n' and 'pathString' annotations in data.tree
Used to convert annotated binary phylogeny tree to ggraph for easier plotting 
1) write data.tree object to Newick using custom Newick function,
2) read Newick into ape tree object
3) tidygraph::as_tbl_graph to convert from ape to ggraph
4) data.tree to node-level dataframe 
5) join node-level dataframe to tbl_graph nodes
The use of ToNewick is a major bottleneck in computation speed right now}
\usage{
data.tree_to_ggraphNW(data.tree, categories, diversities, counts)
}
\description{
data.tree to ggraph conversion, by converting data.tree structure to ape via newick (doesn't carry annotations)
then ape (still doesn't carry annotations) to ggraph, 
then join data frame of data.tree to restore annotations.
requires 'n' and 'pathString' annotations in data.tree
Used to convert annotated binary phylogeny tree to ggraph for easier plotting 
1) write data.tree object to Newick using custom Newick function,
2) read Newick into ape tree object
3) tidygraph::as_tbl_graph to convert from ape to ggraph
4) data.tree to node-level dataframe 
5) join node-level dataframe to tbl_graph nodes
The use of ToNewick is a major bottleneck in computation speed right now
}
