% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARBOLhelper.R
\name{diversityPerIDs}
\alias{diversityPerIDs}
\title{Calculate diversity index for a set of IDs in a dataframe}
\usage{
diversityPerIDs(df, group, diversity_metric = "simpson")
}
\arguments{
\item{df}{a dataframe with group information per cell}

\item{group}{groups column for which to calculate diversity}

\item{diversity_metric}{one of 'shannon', 'simpson', or 'invsimpson'}
}
\value{
a list of diversity per group per cell
}
\description{
Calculate diversity index for a set of IDs in a dataframe
}
\examples{
# in a data tree
node[[sprintf('\%s_diversity',y)]] <- diversityPerIDs(metadata, group=y)
# directly on a dataframe
metadata$condition_diversity <- diversityPerIDs(metadata, group=condition)
}
