% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARBOLhelper.R
\name{efficient_dgcMatrix_aggregator}
\alias{efficient_dgcMatrix_aggregator}
\title{Efficient Aggregation of dgCMatrix Columns}
\usage{
efficient_dgcMatrix_aggregator(mat, groupings, aggregation_function)
}
\arguments{
\item{mat}{A matrix of class "dgCMatrix" to be aggregated.}

\item{groupings}{A numeric vector specifying the group for each row in `mat`.}

\item{aggregation_function}{A function to aggregate the groups. E.g., mean, sum, median.}
}
\value{
A sparse matrix of class "dgCMatrix" with aggregated columns.
}
\description{
This function efficiently aggregates the columns of a sparse matrix (dgCMatrix) based on a given grouping vector.
Various aggregation functions can be applied (e.g., mean, sum, median).
}
\examples{
mat <- Matrix(c(1, 0, 2, 0, 0, 3), nrow = 3, sparse = TRUE)
groups <- c(1, 1, 2)
efficient_dgcMatrix_aggregator(mat, groups, mean)
}
