% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCTieredClustering.R
\name{EnoughDiffExp}
\alias{EnoughDiffExp}
\title{Used in BaseCondition_default, determines if two clusters are different from gene expression}
\usage{
EnoughDiffExp(srobj, ident1, ident2, nUpregulated = 25, nDownregulated = 25)
}
\arguments{
\item{srobj}{seurat object}

\item{ident1}{Idents(srobj) label for group 2}

\item{nUpregulated}{threshold number of genes up regulated for clusters to be called different}

\item{nDownregulated}{threshold number of genes down regulated for clusters to be called different}
}
\value{
bool whether two clusters have differential gene expression between them
}
\description{
Used in BaseCondition_default, determines if two clusters are different from gene expression
}
