% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARBOLhelper.R
\name{feature_zscore}
\alias{feature_zscore}
\title{Compute and add z-scores of feature expression values to srobj@misc$tax_ggraph
REQUIRES srobj@misc$tax_ggraph}
\usage{
feature_zscore(srobj, feature, assay)
}
\arguments{
\item{srobj}{A Seurat object with @misc$tax_ggraph}

\item{feature}{The name of the feature whose expression values are to be z-scored.}

\item{assay}{The name of the assay layer where the data resides. default: RNA}
}
\value{
The updated graph object with the added mean z-scores for the given feature per node.
}
\description{
This function calculates z-scores of the expression values of a given feature
for each cell in the Seurat object. It then adds the mean z-score
to the object's graph representation based on groupings determined by taxonomic trees.
}
