% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCTieredClustering.R
\name{GenTieredClusters}
\alias{GenTieredClusters}
\title{Performs Iterative clustering of a v3 Seurat object}
\usage{
GenTieredClusters(
  srobj,
  cluster_assay = "SCT",
  cells = NULL,
  tier = 0,
  clustN = 0,
  PreProcess_fun = PreProcess_sctransform,
  BaseCondition_fun = BaseCondition_default,
  ChooseOptimalClustering_fun = ChooseOptimalClustering_default,
  saveSROBJdir = NULL,
  figdir = NULL,
  SaveEndNamesDir = NULL,
  SaveEndFileName = NULL
)
}
\arguments{
\item{srobj}{v3 seurat object}

\item{cluster_assay}{assay to use for clustering defaults to "SCT"}

\item{cells}{cellnames if tiered clustering should start on subset of object}

\item{tier}{starting level defaults to 0}

\item{clustN}{cluster starting from default to 0}

\item{PreProcess_fun}{function to use for preproccessing defaults to PreProcess_sctransform}

\item{BaseCondition_fun}{function to determine if clustering should continue to recurse}

\item{ChooseOptimalClustering_fun}{function that returns srobj with clusters in `srobj$Best.Clusters` after choosing optimal clustering resolution}

\item{saveSROBJdir}{where to save seurat objects for each tier and cluster, if null does not save}

\item{figdir}{where to save QC figures for each tier and cluster, if null does not save}

\item{SaveEndFileName}{prefix for all end cluster files}

\item{saveEndNamesDir}{where to save directory of end clusters, if null does not save}
}
\value{
list of lists with all seurat objects (highly recommend using folder arguments for saving outputs)
}
\description{
Iterively clusters seurat object from single cell datasets, choosing optimum resolution parameters at each stage of clustering. Outputs QC plots for each tier and stage.

can return output directories figs/ srobjs/ (see examples) containing subset seurat objects and QC plots matching tiered structure of found in dataset

can also return endclusts/ folder which contains tsv files named with the path to the endcluster (T0C0T1C1.tsv) and contains the cellnames (one per line) that comprise that cluster
}
\examples{
srobj <- readRDS("/path/to/seurat_object.rds")
tiers <- GenTieredclusters(srobj,
                           saveSROBJdir = "~/teiredoutput/srobjs",
                           figdir = "~/teiredoutput/figdir",
                           SaveEndNamesDir = "~/teiredoutput/endclusts")

}
