% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARBOLhelper.R
\name{GetStandardNames}
\alias{GetStandardNames}
\title{Calculate curated names for set of clusters as function of their major type}
\usage{
GetStandardNames(
  srobj,
  figdir,
  max_cells_per_ident = 200,
  celltype_col = "celltype"
)
}
\arguments{
\item{srobj}{a seurat object with a celltype metadata column, specified in celltype_col, and tierNident i.e. srobj@meta.data$tierNident}

\item{figdir}{the figdir from an GenTieredClusters (ARBOL) run}

\item{max_cells_per_ident}{maximum cells to use in FindAllMarkers call. defaults to 200}

\item{celltype_col}{the metadata column corresponding to celltype to assign standard names}
}
\value{
the seurat object with curatedname column in metadata
}
\description{
Calculate curated names for set of clusters as function of their major type
}
\examples{
srobj <- GetStandardNames(srobj,figdir='ARBOLoutput/figs')
}
