% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARBOLhelper.R
\name{LoadTiersAsDF}
\alias{LoadTiersAsDF}
\title{Load 'endclusters' folder from R ARBOL output as a dataframe}
\usage{
LoadTiersAsDF(folder = "./endclusts", maxtiers = 10)
}
\arguments{
\item{folder}{R ARBOL endclusts folder}

\item{maxtiers}{max_tiers parameter used in ARBOL}
}
\value{
dataframe with tier membership per cell including tierNident, the endcluster for that cell. cell barcode column CellID
}
\description{
Load 'endclusters' folder from R ARBOL output as a dataframe
}
\examples{
tiers <- LoadTiersAsDF(folder='./endclusts',maxtiers=10)
#metadata from Seurat object
metadata <- metadata \%>\% left_join(tiers, by='CellID')
}
