% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARBOLhelper.R
\name{makeDataTreeNamesUnique}
\alias{makeDataTreeNamesUnique}
\title{makes data.tree names unique}
\usage{
makeDataTreeNamesUnique(l)
}
\arguments{
\item{l}{A data tree object}
}
\value{
a data tree object with names forced to be unique. Useful when producing node-level dataframes from raw data tree objects
}
\description{
makes data.tree names unique
}
\examples{
makeDataTreeNamesUnique(ARBOLtree)
}
