% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCTieredClustering.R
\name{markers.binom}
\alias{markers.binom}
\title{Binomial signifcance test for on/off gene expression}
\usage{
markers.binom(
  object,
  clust.1,
  clust.2 = NULL,
  effect.size = log(2),
  TPM.mat = NULL,
  Count.mat = NULL
)
}
\arguments{
\item{object}{Seruat object}

\item{clust.1}{label of Ident(object) to use as group 1}

\item{clust.2}{label of Ident(object) to use as group 2 defaults: rest}

\item{effect.size}{log2 foldchange threshold for effect size significance}

\item{TPM.mat}{Transcripts/million matrix}

\item{Count.mat}{Count matrix}
}
\value{
significant gene data.frame
}
\description{
#######################################################################################################
Gratefully stolen from:
S4 class file for Shekhar et al.,
   "Comprehensive classification of retinal bipolar cells using single-cell transcriptomics", Cell, 2016
https://raw.githubusercontent.com/broadinstitute/BipolarCell2016/master/class.R
}
\details{
Ported to work with Seurat v3 by Benjamin Doran Oct 2019
Added columns for FDR and posFrac to output by default

#######################################################################################################
}
