% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARBOLhelper.R
\name{mergeEndclusts}
\alias{mergeEndclusts}
\title{Merges tierNidents with their nearest neighbors in a binary tree if 
their sample diversity and number of cells do not meet specified thresholds.
Directly prunes srobj-attached binary tree. 
We suggest re-calculating tree (or just ggraph for viz) from here using centroidTaxonomy() or remake_ggraph(), 
so that new endclusters are treated as leaf nodes}
\usage{
mergeEndclusts(srobj, sample_diversity_threshold, size_threshold)
}
\arguments{
\item{srobj}{a seurat object with a binarytree calculated in slot srobj@misc$binarytree, 
typically calculated using ARBOLcentroidTaxonomy}

\item{sample_diversity_threshold}{sample diversity below which to prune nodes from tree}

\item{size_threshold}{cluster size below which to prune nodes from tree}
}
\value{
the input seurat object with merged tierNidents in a new metadata column, mergedIdent
}
\description{
Merges tierNidents with their nearest neighbors in a binary tree if 
their sample diversity and number of cells do not meet specified thresholds.
Directly prunes srobj-attached binary tree. 
We suggest re-calculating tree (or just ggraph for viz) from here using centroidTaxonomy() or remake_ggraph(), 
so that new endclusters are treated as leaf nodes
}
\examples{
srobj <- mergeEndclusts(srobj, sample_diversity_threshold = 0.1, size_threshold = 10)
}
