% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARBOLhelper.R
\name{mergeEndclustsCustom}
\alias{mergeEndclustsCustom}
\title{Merges tierNidents with their nearest neighbors in a srobj-attached binary tree by custom thresholds
We suggest re-calculating tree (or just ggraph for viz) from here using centroidTaxonomy() or remake_ggraph(), 
so that new endclusters are treated as leaf nodes}
\usage{
mergeEndclustsCustom(srobj, threshold_attributes, thresholds)
}
\arguments{
\item{srobj}{a seurat object with a binarytree calculated in layer srobj@misc$taxTree, typically calculated using ARBOLcentroidTaxonomy}

\item{threshold_attributes}{list of srobj metadata columns to threshold on}

\item{thresholds}{list of threshold values to prune, in same order as threshold_attributes}
}
\value{
the input seurat object with merged tierNidents in a new metadata column, mergedIdent 
and a merged data.tree object in srobj@misc$workingTree
}
\description{
Merges tierNidents with their nearest neighbors in a srobj-attached binary tree by custom thresholds
We suggest re-calculating tree (or just ggraph for viz) from here using centroidTaxonomy() or remake_ggraph(), 
so that new endclusters are treated as leaf nodes
}
\examples{
srobj <- mergeEndclustsCustom(srobj, threshold_attributes = c('sample_diversity','n'), 
                              thresholds = c(0.2,50))
}
