% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARBOLhelper.R
\name{mergeEndclustsCustomIdents}
\alias{mergeEndclustsCustomIdents}
\title{Merges tierNidents with their nearest neighbors in a srobj-attached binary tree by custom thresholds, 
outputs dataframe of new idents}
\usage{
mergeEndclustsCustomIdents(srobj, threshold_attributes, thresholds)
}
\arguments{
\item{srobj}{a seurat object with a binarytree calculated in slot srobj@misc$taxTree, typically calculated using ARBOLcentroidTaxonomy}

\item{threshold_attributes}{list of srobj metadata columns to threshold on}

\item{thresholds}{list of threshold values to prune, in same order as threshold_attributes}
}
\value{
dataframe of new idents
}
\description{
Merges tierNidents with their nearest neighbors in a srobj-attached binary tree by custom thresholds, 
outputs dataframe of new idents
}
\examples{
mIdents <- mergeEndclustsCustomIdents(srobj, threshold_attributes = c('sample_diversity','n'), thresholds = c(0.2,50))
}
