% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARBOLhelper.R
\name{mergeEndclustsIdents}
\alias{mergeEndclustsIdents}
\title{Merge based on sample diversity and endclust size, only outputting new end-clusters per cell}
\usage{
mergeEndclustsIdents(srobj, sample_diversity_threshold, size_threshold)
}
\arguments{
\item{srobj}{a seurat object with a binarytree calculated in slot srobj@misc$binarytree, typically calculated using ARBOLcentroidTaxonomy}

\item{sample_diversity_threshold}{sample diversity below which to prune nodes from tree}

\item{size_threshold}{cluster size below which to prune nodes from tree}
}
\value{
dataframe with 2 columns, cellid and new endcluster
}
\description{
Merge based on sample diversity and endclust size, only outputting new end-clusters per cell
}
\examples{
srobj <- mergeEndclustsIdents(srobj, sample_diversity_threshold = 0.1, size_threshold = 10)
}
