% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARBOLhelper.R
\name{pieify_tree_plot}
\alias{pieify_tree_plot}
\title{Transform ggraph tree plot to include pie charts at nodes}
\usage{
pieify_tree_plot(
  ggraph_plot,
  srobj,
  color_metadata,
  y_cutoff = 1,
  radius = 0.5,
  mode = "subclusters",
  scaling_factor = 1
)
}
\arguments{
\item{ggraph_plot}{ggraph tree plot with at least one geom_node_point() layer}

\item{srobj}{Seurat object with srobj@misc$tax_ggraph that has metadata column for pie-chart graphs}

\item{color_metadata}{variable for which to chart pies}

\item{y_cutoff}{height on the chart above which to chart pies}

\item{radius}{pie radius}

\item{mode}{type of tree plot to add piecharts to - one of "subclusters" or "taxonomy"}

\item{scaling_factor}{scaling factor used to change pie sizes in "subclusters" mode}
}
\value{
ggraph tree plot with pie charts at each node above the y_cutoff height
}
\description{
Transform ggraph tree plot to include pie charts at nodes
}
\examples{
b <- ggraph(srobj@misc$tax_ggraph, layout='dendrogram', height=plotHeight*20) + 
       geom_edge_elbow2(aes(color=node.subset_majority),edge_width=1) +
   new_scale_color() +
   geom_node_text(aes(filter = leaf, label = subset_majority \%>\% str_replace_all('_','/'), color = subset_majority), 
                  nudge_y=-2.4,vjust=0.5,hjust=0,angle=270,size=8) +
   scale_color_manual(values=pie.colors) +
   new_scale_color() +
   geom_node_text(aes(filter = leaf, label = n),color='grey30',nudge_y=-0.2,vjust=0.5,hjust=0,size=8,angle=270) + 
   theme_classic() +
   geom_node_point(aes(filter = leaf, color=sample_diversity),size=4,shape='square') + 
   scale_color_gradient(low='grey90',high='grey10') +
   expand_limits(y=-5) + 
   coord_cartesian(xlim=c(0,35),ylim=c(-10,10)) +
   geom_node_point(aes(label=string),size=0)
pietree_plot <- pieify_tree_plot(b, srobj, 'polyp', 'Set1',mode='taxonomy')
}
