% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARBOL.R
\name{Plotting}
\alias{Plotting}
\title{save basic cell counts
basic umap of clusters (using umap-learn, as only it is allowed on premade graphs)
output heatmap of markers}
\usage{
Plotting(working_srobj, fig_dir = NULL)
}
\description{
save basic cell counts
basic umap of clusters (using umap-learn, as only it is allowed on premade graphs)
output heatmap of markers
}
