% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARBOLhelper.R
\name{preppedTree_toDF}
\alias{preppedTree_toDF}
\title{Converts data.tree object to dataframe with row per node, this function aids conversion of highly annotated data.tree objects to dataframes.
Used in ARBOL workflow to create a tbl_graph object, a class useful for plotting dendrograms}
\usage{
preppedTree_toDF(tree, ...)
}
\arguments{
\item{tree}{usually a custom annotated data.tree object}
}
\value{
a dataframe of the tree with one row per node
}
\description{
Converts data.tree object to dataframe with row per node, this function aids conversion of highly annotated data.tree objects to dataframes.
Used in ARBOL workflow to create a tbl_graph object, a class useful for plotting dendrograms
}
\examples{
ARBOLdf <- preppedTree_toDF(divtree)
txt <- ToNewick(divtree)
ARBOLphylo <- ape::read.tree(text=txt)

x <- as_tbl_graph(ARBOLphylo, directed=T) \%>\% activate(nodes) \%>\% 
left_join(ARBOLdf \%>\% select(name=pathString,sample_diversity,disease_diversity))
}
