% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARBOL.R
\name{PreProcess_no_normalization}
\alias{PreProcess_no_normalization}
\title{seurat v3 processing}
\usage{
PreProcess_no_normalization(
  srobj,
  ChoosePCs_fun = ChoosePCs_default,
  ChooseHVG_fun = ChooseHVG_default,
  fig_dir = NULL,
  ...
)
}
\description{
seurat v3 processing
}
