% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARBOL.R, R/arboldotdotdot.R
\name{PreProcess_sctransform_harmony}
\alias{PreProcess_sctransform_harmony}
\title{seurat v4 processing}
\usage{
PreProcess_sctransform_harmony(srobj, fig_dir = NULL, regressVar = NULL, ...)

PreProcess_sctransform_harmony(srobj, fig_dir = NULL, regressVar = NULL, ...)
}
\description{
seurat v4 processing

seurat v4 processing
}
