% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARBOLhelper.R
\name{prepSubclusteringMetadata}
\alias{prepSubclusteringMetadata}
\title{Prepare ARBOL seurat object metadata for tree building}
\usage{
prepSubclusteringMetadata(
  srobj,
  maxtiers = 10,
  categorical_attributes,
  diversity_attributes,
  numerical_attributes,
  total_attributes
)
}
\arguments{
\item{srobj}{a seurat object with ARBOL 'tierNident' and 'sample' columns}

\item{maxtiers}{max_tiers parameter used in ARBOL}

\item{categorical_attributes}{categorical attributes are propagated up a tree as unique values per node. also calculates a majority per node
majority added as node$attribute_majority}

\item{diversity_attributes}{attributes for which to calculate diversity in each node.}

\item{numerical_attributes}{numerical attributes are propagated up a tree as counts of each unique value per node.
creates a new column per unique value per attribute. counts are added as node$attribute_n_value}

\item{total_attributes}{total attributes are numerical attributes that you want to propagate up the tree by summing
the value of each node.
Supports simpson's, inverse simpson's, and shannon index as implemented in vegan diversity()}
}
\value{
a metadata dataframe ready for tree building
}
\description{
Prepare ARBOL seurat object metadata for tree building
}
\examples{
prepSubclusteringMetadata(srobj, maxtiers=10)
}
