% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARBOLhelper.R
\name{prepTree}
\alias{prepTree}
\title{Prepare ARBOL seurat object metadata for tree building}
\usage{
prepTree(
  ARBOLtree,
  srobj,
  numerical_attributes,
  categorical_attributes,
  diversity_attributes = "sample"
)
}
\arguments{
\item{srobj}{a seurat object with ARBOL 'tierNident' and 'sample' columns}

\item{maxtiers}{max_tiers parameter used in ARBOL}
}
\value{
a metadata dataframe ready for tree building
}
\description{
Prepare ARBOL seurat object metadata for tree building
}
\examples{
prepARBOLmeta_tree(srobj, maxtiers=10)
}
