% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARBOL.R, R/arboldotdotdot.R
\name{QC_Plotting}
\alias{QC_Plotting}
\title{basic QC scatter plots}
\usage{
QC_Plotting(working_srobj, fig_dir = NULL, ...)

QC_Plotting(working_srobj, fig_dir = NULL, ...)
}
\description{
basic QC scatter plots

basic QC scatter plots
}
