% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARBOLhelper.R
\name{remake_ggraph}
\alias{remake_ggraph}
\title{remake ggraph object with new categories and diversities, useful to add curatednames}
\usage{
remake_ggraph(
  srobj,
  categories,
  diversities,
  counts,
  diversity_metric = "simpson"
)
}
\arguments{
\item{srobj}{a seurat object with tierNident, sample, and category columns in metadata i.e. srobj@meta.data$tierNident}

\item{categories}{categories as in tree building functions}

\item{diversities}{diversities as in tree building functions}

\item{counts}{attributes for which to count unique values per node.}

\item{diversity_metric}{one of 'shannon', 'simpson', or 'invsimpson'}
}
\value{
the seurat object with ggraph object remade in srobj@misc$binarytreeggraph
}
\description{
remake ggraph object with new categories and diversities, useful to add curatednames
}
\examples{
srobj <- remake_ggraph(srobj, categories = c('curatedname','celltype'), diversities = c('curatedname','celltype'))
}
