% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARBOLhelper.R
\name{SIperGroup}
\alias{SIperGroup}
\title{Calculate Simpson's index per group in a dataframe (only creates 'diversity column', can't do multiple kinds),}
\usage{
SIperGroup(df, species, group)
}
\arguments{
\item{df}{a dataframe with species and groups columns}

\item{species}{species column to stratify per group}

\item{group}{groups column for which to calculate diversity}
}
\value{
dataframe with species and diversity columns
}
\description{
Calculate Simpson's index per group in a dataframe (only creates 'diversity column', can't do multiple kinds),
}
\examples{
dataframe <- dataframe \%>\% left_join(SIperGroup(dataframe, species=pathString, group='sample')) \%>\% 
                           suppressMessages
metadata <- metadata \%>\% left_join(SIperGroup(metadata, species=tierNident, group='sample')) \%>\% 
                         suppressMessages
}
