% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARBOLhelper.R
\name{SIperIDs}
\alias{SIperIDs}
\title{Calculate Simpson's index for a set of IDs in a dataframe}
\usage{
SIperIDs(df, group)
}
\arguments{
\item{df}{a dataframe with group information per cell}

\item{group}{groups column for which to calculate diversity}
}
\value{
a list of diversity per group per cell
}
\description{
Calculate Simpson's index for a set of IDs in a dataframe
}
\examples{
in a data tree
node[[sprintf('\%s_diversity',y)]] <- SIperIDs(metadata, group=y)
directly on a dataframe
metadata$condition_diversity <- SIperIDs(metadata, group=condition)
}
