% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCTieredClustering.R
\name{SplitSrobjOnIdents}
\alias{SplitSrobjOnIdents}
\title{turns single seurat object into list of seurat objects, one for each ident in Idents(srobj)}
\usage{
SplitSrobjOnIdents(srobj, proj_nm_suffix)
}
\arguments{
\item{srobj}{seurat object}

\item{project_nm_suffix}{suffix to append to `srobj@project.name`}
}
\value{
list of seurat objects
}
\description{
turns single seurat object into list of seurat objects, one for each ident in Idents(srobj)
}
