% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARBOLhelper.R
\name{spread_tierN}
\alias{spread_tierN}
\title{spread tierNident column in a dataframe into separate columns per tier, with full paths and single tier idents}
\usage{
spread_tierN(df, max_tiers = 10, sep = ".")
}
\arguments{
\item{df}{a dataframe with ARBOL tierNident (usually srobj@meta.data)}

\item{max_tiers}{max_tiers parameter used in ARBOL}
}
\value{
the dataframe with separate tier columns, useful for producing and plotting dendrograms
}
\description{
spread tierNident column in a dataframe into separate columns per tier, with full paths and single tier idents
}
\examples{
srobj@meta.data <- spread_tierN(srobj@meta.data)
}
