% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARBOLhelper.R
\name{sr_ARBOLclustertree}
\alias{sr_ARBOLclustertree}
\title{Creates data tree object for ARBOL run, adds it to seurat object along with a graph of it
calls prepARBOLmeta_tree() and prepTree()}
\usage{
sr_ARBOLclustertree(srobj)
}
\arguments{
\item{srobj}{a seurat object with ARBOL 'tierNident' and 'sample' columns.}
}
\value{
the input seurat object with tiered clustering tree in srobj@misc$pvclust and ggraph of tree to srobj@misc$ARBOLclustertreegraph
}
\description{
Creates data tree object for ARBOL run, adds it to seurat object along with a graph of it
calls prepARBOLmeta_tree() and prepTree()
}
\examples{
srobj <- sr_ARBOLclustertree(srobj)
}
