% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARBOLhelper.R
\name{sr_binarytree}
\alias{sr_binarytree}
\title{Calculate pvclust() tree (a binary tree of distances between end-clusters) for ARBOL results
tree based on euclidean distance between cluster centroids based on gene medians with complete linkage}
\usage{
sr_binarytree(srobj, assay = "SCT")
}
\arguments{
\item{srobj}{a seurat object with ARBOL 'tierNident' column}

\item{assay}{the seurat object assay in which to calculate gene medians. Defaults to 'SCT'}
}
\value{
the input seurat object with pvclust tree in srobj@misc$pvclust
}
\description{
Calculate pvclust() tree (a binary tree of distances between end-clusters) for ARBOL results
tree based on euclidean distance between cluster centroids based on gene medians with complete linkage
}
\examples{
srobj <- sr_binarytree(srobj)
}
