% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARBOLhelper.R
\name{subclusteringTree}
\alias{subclusteringTree}
\title{Creates data tree object for ARBOL run, adds it to seurat object along with a graph of it
calls prepSubclusteringMetadata() and propagateTree()}
\usage{
subclusteringTree(
  srobj,
  categories = "sample",
  diversities = "sample",
  diversity_metric = "simpson",
  counts = "sample",
  totals = "nCount_RNA"
)
}
\arguments{
\item{srobj}{a seurat object with ARBOL 'tierNident', 'CellID', and 'sample' columns.}

\item{categories}{columns in metadata for which you want to track categories present per node. Also finds the majority per node}

\item{diversities}{columns in metadata for which you want to calculate diversity per node}

\item{diversity_metric}{one of 'shannon', 'simpson', or 'invsimpson'}

\item{counts}{columns in metadata for which you want to count each value per node}

\item{totals}{attributes for which to sum values per node.}
}
\value{
the input seurat object with tiered clustering tree in srobj@misc$subclusteringTree, 
plot of tree to srobj@misc$subclusteringViz, and ggraph object to srobj@misc$cluster_ggraph
}
\description{
Creates data tree object for ARBOL run, adds it to seurat object along with a graph of it
calls prepSubclusteringMetadata() and propagateTree()
}
\examples{
srobj <- subclusteringTree(srobj)
}
