% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARBOLhelper.R
\name{sv.cbind}
\alias{sv.cbind}
\title{Combine Sparse Vectors into a Sparse Matrix
from https://stackoverflow.com/questions/8843700/creating-sparse-matrix-from-a-list-of-sparse-vectors
This function takes a list of sparse vectors and combines them into a single sparse matrix.
The vectors in the list must all have the same length.}
\usage{
sv.cbind(input_list)
}
\arguments{
\item{input_list}{List of sparse vectors of class "dsparseVector".}
}
\value{
A sparse matrix of class "dgCMatrix".
}
\description{
Combine Sparse Vectors into a Sparse Matrix
from https://stackoverflow.com/questions/8843700/creating-sparse-matrix-from-a-list-of-sparse-vectors
This function takes a list of sparse vectors and combines them into a single sparse matrix.
The vectors in the list must all have the same length.
}
\examples{
vec1 <- sparseVector(c(1, 2), c(1, 2))
vec2 <- sparseVector(c(3), c(2))
sv.cbind(list(vec1, vec2))
}
