% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARBOLhelper.R
\name{tierN_diversity}
\alias{tierN_diversity}
\title{Calculate Diversity Index for tierNidents for multiple attributes of the data
diversity metric options same as those available in vegan's diversity() function}
\usage{
tierN_diversity(srobj, diversity_attributes, diversity_metric = "simpson")
}
\arguments{
\item{srobj}{a seurat object with ARBOL tierNident column}

\item{diversity_attributes}{the attributes you wish to calculate diversity for (e.g. disease, sample)}

\item{diversity_metric}{one of 'shannon', 'simpson', or 'invsimpson'}
}
\value{
the seurat object with diversity per attribute per tierNident added to the metadata
}
\description{
Calculate Diversity Index for tierNidents for multiple attributes of the data
diversity metric options same as those available in vegan's diversity() function
}
\examples{
srobj <- tierN_diversity(srobj, diversity_attributes = c('sample','disease'))
}
