% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARBOLhelper.R
\name{tierN_SI}
\alias{tierN_SI}
\title{Calculate Gini-Simpson's Index for tierNidents for multiple attributes of the data}
\usage{
tierN_SI(srobj, diversity_attributes)
}
\arguments{
\item{srobj}{a seurat object with ARBOL tierNident column}

\item{diversity_attributes}{the attributes you wish to calculate diversity for (e.g. disease, sample)}
}
\value{
the seurat object with diversity per attribute per tierNident added to the metadata
}
\description{
Calculate Gini-Simpson's Index for tierNidents for multiple attributes of the data
}
\examples{
srobj <- tierN_SI(srobj, diversity_attributes = c('sample','disease'))
}
