% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARBOLhelper.R
\name{ToNewickPS}
\alias{ToNewickPS}
\title{data.tree to phyloseq object conversion with custom ToNewick function that uses pathString as node names.}
\usage{
ToNewickPS(node, heightAttribute = DefaultPlotHeight, ...)
}
\arguments{
\item{node}{A data tree object}
}
\value{
Newick format tree
}
\description{
data.tree to phyloseq object conversion with custom ToNewick function that uses pathString as node names.
}
\examples{
txt <- ToNewickPS(divtree)
ARBOLphylo <- ape::read.tree(text=txt)
}
