% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARBOLhelper.R
\name{treeAllotment}
\alias{treeAllotment}
\title{Allots seurat object metadata attributes to a data.tree object
Once attributes are alloted, propagateTree should be run to propagate attributes up the tree.}
\usage{
treeAllotment(srobj, treedf, categories, diversities, diversity_metric, counts)
}
\arguments{
\item{srobj}{a seurat object with ARBOL 'tierNident', 'CellID', 'sample' columns.}

\item{categories}{categorical attributes are propagated up a tree as unique values per node. also calculates a majority per node}

\item{diversities}{attributes for which to calculate diversity in each node. Currently only supports gini-simpson's index.}

\item{diversity_metric}{one of 'shannon', 'simpson', or 'invsimpson'}

\item{counts}{attributes for which to count unique values per node.}
}
\value{
tree with attributes alloted to internal nodes
}
\description{
Allots seurat object metadata attributes to a data.tree object
Once attributes are alloted, propagateTree should be run to propagate attributes up the tree.
}
