% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCTieredClustering.R
\name{ChooseOptimalClustering_default}
\alias{ChooseOptimalClustering_default}
\title{wrapper for ChooseClusterResolutionDownsample enables downsampling srobj before choosing resolution}
\usage{
ChooseOptimalClustering_default(
  srobj,
  downsample_num = Inf,
  PreProcess_fun = PreProcess,
  ...
)
}
\arguments{
\item{srobj}{seurat object}

\item{downsample_num}{max number of cells to allow in seurat object, downsamples to this number}

\item{PreProcess_fun}{function to use to preprocess seurat object, must make ready for calling Seurat::FindClusters}

\item{...}{optional arguments for ChooseClusterResolutionDownsample}
}
\value{
seurat object with optimal cluster in `srobj$Best.Clusters` & resolution choice in `srobj@misc$resolution.choice`
}
\description{
wrapper for ChooseClusterResolutionDownsample enables downsampling srobj before choosing resolution
}
