% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCTieredClustering.R
\name{ChoosePCs_default}
\alias{ChoosePCs_default}
\title{if num cells > 500 use elbow plot if small use jackstraw, 
  if none sig use first 2 pcs}
\usage{
ChoosePCs_default(
  srobj,
  improved_diff_quantile = 0.85,
  significance = 0.01,
  fig_dir = NULL
)
}
\description{
if num cells > 500 use elbow plot if small use jackstraw, 
  if none sig use first 2 pcs
}
