% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARBOLhelper.R
\name{data.tree_to_ggraph}
\alias{data.tree_to_ggraph}
\title{data.tree to ggraph conversion, by converting data.tree structure (doesn't carry annotations) to ggraph, 
then join data frame of data.tree
requires 'n' and 'pathString' annotations in data.tree
Used to convert annotated binary phylogeny tree to ggraph for easier plotting 
1) write data.tree object to Newick using custom Newick function,
2) read Newick into ape tree object
3) ggraph::as_tbl_graph to convert from ape to ggraph
4) data.tree to node-level dataframe 
5) join node-level dataframe to tbl_graph nodes}
\usage{
data.tree_to_ggraph(data.tree, categories, diversities)
}
\description{
data.tree to ggraph conversion, by converting data.tree structure (doesn't carry annotations) to ggraph, 
then join data frame of data.tree
requires 'n' and 'pathString' annotations in data.tree
Used to convert annotated binary phylogeny tree to ggraph for easier plotting 
1) write data.tree object to Newick using custom Newick function,
2) read Newick into ape tree object
3) ggraph::as_tbl_graph to convert from ape to ggraph
4) data.tree to node-level dataframe 
5) join node-level dataframe to tbl_graph nodes
}
