% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARBOLhelper.R
\name{data.tree_to_ggraph}
\alias{data.tree_to_ggraph}
\title{data.tree to ggraph conversion, by converting data.tree structure to dendrogram via custom PS function
then ape (still doesn't carry annotations) to ggraph,
then join data frame of data.tree to restore annotations.
requires 'n' and 'pathString' annotations in data.tree
Used to convert annotated binary phylogeny tree to ggraph for easier plotting 
1) write data.tree object to dend object using custom as.dendrogram.Node function,
2) convert dendrogram structure object to ggraph using tidygraph::as_tbl_graph
3) left join heights, categories, diversities, and counts to tbl_graph object
previous version data.tree_to_ggraphNW used highly inefficient newick text conversion for tree structure}
\usage{
data.tree_to_ggraph(
  data.tree,
  categories,
  diversities,
  counts,
  totals,
  heightAttribute = "plotHeight"
)
}
\description{
data.tree to ggraph conversion, by converting data.tree structure to dendrogram via custom PS function
then ape (still doesn't carry annotations) to ggraph,
then join data frame of data.tree to restore annotations.
requires 'n' and 'pathString' annotations in data.tree
Used to convert annotated binary phylogeny tree to ggraph for easier plotting 
1) write data.tree object to dend object using custom as.dendrogram.Node function,
2) convert dendrogram structure object to ggraph using tidygraph::as_tbl_graph
3) left join heights, categories, diversities, and counts to tbl_graph object
previous version data.tree_to_ggraphNW used highly inefficient newick text conversion for tree structure
}
