% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARBOLhelper.R
\name{DiversityPerGroup}
\alias{DiversityPerGroup}
\title{Calculate diversity index per group in a dataframe (adds '%s_diversity' column),
diversity metric options same as those available in vegan diversity() function}
\usage{
DiversityPerGroup(df, species, group, diversity_metric = "simpson")
}
\arguments{
\item{df}{a dataframe with species and groups columns}

\item{species}{species column to stratify per group}

\item{group}{groups column for which to calculate diversity}

\item{diversity_metric}{one of 'shannon', 'simpson', or 'invsimpson'}
}
\value{
dataframe with species and diversity columns
}
\description{
Calculate diversity index per group in a dataframe (adds '%s_diversity' column),
diversity metric options same as those available in vegan diversity() function
}
\examples{
dataframe <- dataframe \%>\% left_join(DiversityPerGroup(dataframe, species=pathString, group='sample')) \%>\% 
                           suppressMessages
metadata <- metadata \%>\% left_join(DiversityPerGroup(metadata, species=tierNident, group='sample')) \%>\% 
                         suppressMessages
}
