% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARBOLhelper.R
\name{MergeEndclusts}
\alias{MergeEndclusts}
\title{Merges tierNidents with their nearest neighbors in a binary tree if their sample diversity and number of cells do not meet thresholds}
\usage{
MergeEndclusts(srobj, sample_diversity_threshold, size_threshold)
}
\arguments{
\item{srobj}{a seurat object with a binarytree calculated in slot srobj@misc$binarytree, typically calculated using sr_ARBOLbinarytree}
}
\value{
the input seurat object with merged tierNidents in a new metadata column, mergedIdent
}
\description{
Merges tierNidents with their nearest neighbors in a binary tree if their sample diversity and number of cells do not meet thresholds
}
\examples{
srobj <- MergeEndclusts(srobj, sample_diversity_threshold = 0.1, size_threshold = 10)
}
