% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARBOLhelper.R
\name{outputStandardNames}
\alias{outputStandardNames}
\title{Calculate curated names for set of clusters as function of their major type}
\usage{
outputStandardNames(
  srobj,
  figdir,
  max_cells_per_ident = 200,
  celltype_col = "celltype",
  standardname_col = "curatedname",
  n_genes = 2
)
}
\arguments{
\item{srobj}{a seurat object with a celltype metadata column, specified in celltype_col, 
CellID (for preserving metadata rownames), and tierNident (i.e. srobj@meta.data$CellID, srobj@meta.data$tierNident)}

\item{figdir}{the figdir from an ARBOL run}

\item{max_cells_per_ident}{maximum cells to use in FindAllMarkers call. defaults to 200}

\item{celltype_col}{the metadata column corresponding to celltype to assign standard names}

\item{standardname_col}{metadata column to which to output celltype names. defaults to 'curatedname'}

\item{n_genes}{number of genes with which to create standard names. defaults to 2}
}
\value{
a dataframe with standard names per end-cluster
}
\description{
Calculate curated names for set of clusters as function of their major type
}
\examples{
srobj <- outputStandardNames(srobj,figdir='ARBOLoutput/figs')
}
