% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARBOL.R
\name{PreProcess}
\alias{PreProcess}
\title{seurat v3 processing}
\usage{
PreProcess(
  srobj,
  ChoosePCs_fun = ChoosePCs_default,
  ChooseHVG_fun = ChooseHVG_default,
  fig_dir = NULL,
  ...
)
}
\description{
seurat v3 processing
}
