% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARBOL.R
\name{PreProcess_sctransform}
\alias{PreProcess_sctransform}
\title{seurat v4 processing}
\usage{
PreProcess_sctransform(
  srobj,
  ChoosePCs_fun = ChoosePCs_default,
  fig_dir = NULL,
  ...
)
}
\description{
seurat v4 processing
}
