% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARBOLhelper.R
\name{sr_ARBOLbinarytree}
\alias{sr_ARBOLbinarytree}
\title{Creates binary tree object for ARBOL run, adds it to seurat object along with a graph of it
calls sr_binarytree() in which assay + methods for tree building are called}
\usage{
sr_ARBOLbinarytree(srobj, diversity_attributes = "sample")
}
\arguments{
\item{srobj}{a seurat object with ARBOL 'tierNident', 'CellID', 'sample' columns.}
}
\value{
the input seurat object with binary tree pvclust object in srobj@misc$pvclust, 
bina
}
\description{
Creates binary tree object for ARBOL run, adds it to seurat object along with a graph of it
calls sr_binarytree() in which assay + methods for tree building are called
}
\examples{
srobj <- sr_ARBOLbinarytree(srobj)
}
