% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARBOLhelper.R
\name{sr_ARBOLclustertree}
\alias{sr_ARBOLclustertree}
\title{Creates data tree object for ARBOL run, adds it to seurat object along with a graph of it
calls prepARBOLmeta_tree() and prepTree()}
\usage{
sr_ARBOLclustertree(
  srobj,
  categories = "sample",
  diversities = "sample",
  diversity_metric = "simpson",
  counts = NA
)
}
\arguments{
\item{srobj}{a seurat object with ARBOL 'tierNident', 'CellID', and 'sample' columns.}

\item{categories}{columns in metadata for which you want to track categories present per node. Also finds the majority per node}

\item{diversities}{columns in metadata for which you want to calculate diversity per node}

\item{diversity_metric}{one of 'shannon', 'simpson', or 'invsimpson'}

\item{counts}{columns in metadata for which you want to count each value per node}
}
\value{
the input seurat object with tiered clustering tree in srobj@misc$ARBOLclustertree, 
plot of tree to srobj@misc$ARBOLclustertreeviz, and ggraph object to srobj@misc$ARBOLclustertreeggraph
}
\description{
Creates data tree object for ARBOL run, adds it to seurat object along with a graph of it
calls prepARBOLmeta_tree() and prepTree()
}
\examples{
srobj <- sr_ARBOLclustertree(srobj)
}
