% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARBOLhelper.R
\name{sr_binarytree}
\alias{sr_binarytree}
\title{Calculate pvclust() tree (a binary tree of distances between end-clusters) for ARBOL results
tree based on euclidean distance between cluster centroids based on gene medians with complete linkage}
\usage{
sr_binarytree(
  srobj,
  tree_reduction = "centroids",
  hclust_method = "complete",
  distance_method = "euclidean",
  centroid_method = "mean",
  centroid_assay = "SCT",
  reduction_dims = 1:25
)
}
\arguments{
\item{srobj}{a seurat object with ARBOL 'tierNident' column}

\item{tree_reduction}{either 'centroids', which calculates centroids among all genes, or any reduction slot in srobj}

\item{hclust_method}{any hierarchical clustering method implemented in pvclust::pvclust(method.hclust), defaults to 'complete'}

\item{distance_method}{any distance method implemented in pvclust::pvclust(method.dist) - one of "correlation", "abscor", "uncentered", "euclidean" -
or cosine (no quotes) as implemented in ARBOL::cosine. you may also write your own function that returns a dist object, as in pvclust::pvclust()}

\item{centroid_method}{the function used to calculate centroids in the tree_reduction matrix, as implemented in Matrix.utils::aggregate.Matrix(fun)
currently, sum, count, mean, and median are supported}

\item{centroid_assay}{if using cell x gene data (not any srobj@reduction), the assay within which to calculate centroids}

\item{reduction_dims}{the dimensions of the reduction slot to use for centroid calculation. defaults to 1:25}
}
\value{
the input seurat object with pvclust tree in srobj@misc$pvclust
}
\description{
Calculate pvclust() tree (a binary tree of distances between end-clusters) for ARBOL results
tree based on euclidean distance between cluster centroids based on gene medians with complete linkage
}
\examples{
srobj <- sr_binarytree(srobj = srobj, tree_reduction = 'pca', hclust_method = 'complete',
                         distance_method = 'euclidean', centroid_method = 'median', 
                         centroid_assay = 'SCT', reduction_dims = 1:25)
}
